Add bFrameAdjustment configuration option to address PTS/DTS bug
WIP: pending discussion, will add/fix tests

This change introduces a configuration option that adjusts
the timestampOffset and appendWindowEnd to address the PTS/DTS
bug in Chrome. It assumes that the number of bframes is known,
and constant. It also assumes that the manifest provides an
accurate framerate.

See the full writeup in externs/shaka/player.js